+++ packages/blocks/src/root-block/widgets/index.ts
@@ -1,3 +1,8 @@
+export {
+  AFFINE_AI_ACTION_PANEL_WIDGET,
+  AffineAIActionPanelWidget,
+  type AffineAIActionPanelWidgetConfig,
+} from './ai-action-panel/ai-action-panel.js';
 export { BlockHubWidget } from './block-hub/block-hub.js';
 export { AffineDocRemoteSelectionWidget } from './doc-remote-selection/doc-remote-selection.js';
 export { AffineDragHandleWidget } from './drag-handle/drag-handle.js';
